[border-agent] add 'rv' key to MeshCoP TXT data (#11504)
This commit updates `BorderAgent::PrepareServiceTxtData()` to include
the "rv" key. This key represents the version of the TXT record
format. Per the Thread specification, it must be set to "1". Values
other than "1" are reserved for the future and MUST NOT be used.

This commit also updates `test_border_agent` to validate this key.