[diag] deprecate otPlatDiagRadioTransmitDone() and otPlatDiagRadioReceiveDone()
The radio platform API otPlatDiagRadioTransmitDone() and otPlatDiagRadioReceiveDone()
are the same with the API otPlatRadioTxDone() and otPlatRadioReceiveDone(). This commit
removes the API otPlatDiagRadioTransmitDone() and otPlatDiagRadioReceiveDone() to let
the MAC layer and the diag module to use the same radio API to send and receive 154
frames. So that the diag module could process the ACK frame in the future.