[dhcp6-pd-client] initialize `msgType` to fix compiler warning
This commit updates `SendMessage()` to initialize the `msgType`
variable before the `switch` statement.

This change addresses a compiler warning for a possibly uninitialized
variable, flagged by `-Werror=maybe-uninitialized`.

Note that the situation where `mState` would be an undefined value is
not technically possible in the current logic. However, the compiler
cannot guarantee this and therefore generates a warning. Initializing
the variable upfront resolves this issue.