[mle] register `Message::TxCallback` when sending "Child ID Req" (#11720)
This commit updates the mechanism for notifying the transmit
completion of a Child ID Request message.

Instead of `MeshForwarder` checking the message type and calling into
`Mle` upon transmit completion, `Mle::SendChildIdRequest()` now
registers a `TxCallback` directly on the message.

This change simplifies `MeshForwarder` by removing the need for it to
be aware of specific `Mle` message types.