[mle] reorganize `Detacher` methods
This commit moves the `Detacher::Detach()` method implementation to
group it with other `Detacher` method definitions.

When the `Detacher` class was added in PR #11723, its methods were
intentionally kept in their original locations to keep the `git diff`
smaller and easier to review.

This is a pure code-move refactoring to improve code organization and
has no functional changes.