[socket] Fix default netif used when opening an UDP socket
After kNetifThreadHost was introduced, the default netif for the
UDP open API was changed from kNetifThread to kNetifThreadHost.
This change can cause issues on devices that implement the
UDP platform code.
For example, if the Thread shell is used to open a socket and then
connect is called to set the destination to another device,
the transmission may fail in some cases. This happens because
the packet is sent via the host Thread interface instead of
the internal one. The default value should be the internal interface,
and if the user needs to change it, they can use the bind command.

The default interface has also been updated to kNetifThreadInternal
for the DNS and SNP clients, as these modules are intended for
communication within the Thread network.

Signed-off-by: Marius Preda <marius.preda@nxp.com>