[tmf] add `SendOwnedMessage()` to simplify message management
This commit introduces `CoapBase::SendOwnedMessage()` to simplify
message buffer management for callers.

The new method accepts an `OwnedPtr<Message>`, allowing callers to
transfer ownership of a message to the TMF agent upon a successful
send.

This change eliminates the common pattern of using
`FreeMessageOnError` and makes the message management much safer and
simpler. All callers that create and send TMF messages are updated to
use the new `OwnedPtr` pattern.