[mle] move child ID allocation logic to `ChildTable`
This commit moves the child ID allocation logic from `Mle` to
`ChildTable`.

A new `AllocateNewChildRloc16()` method is added to `ChildTable` to
contain the allocation logic, and the `mNextChildId` counter is moved
to `ChildTable`. `Mle` is updated to use this new method.

This refactoring improves code encapsulation by making the
`ChildTable` responsible for managing all aspects of the children it
contains, including ID allocation.