[ip6] introduce `Ip6::RxMessage` to encapsulate `MessageInfo`
This change introduces `Ip6::RxMessage` which inherits from
`ot::Message` and is intended to be used for a received IPv6 packet.

`Ip6::RxMessage` encapsulates the `Ip6::MessageInfo` associated with
the received message. This is made possible by adding a new `void*`
info member to `ot::Message` which `RxMessage` uses to store a
pointer to its `MessageInfo`.

The main goal is to simplify the next layer IPv6/UDP handlers. The UDP
receive handler signature is changed from `(Message &aMessage, const
MessageInfo &aMessageInfo)` to a simpler `
(Ip6::RxMessage &aMessage)`.

All UDP receive handlers are updated to use the new `Ip6::RxMessage`.