[history-tracker] add support for tracking favored on-link prefix
This commit adds support to `HistoryTracker` to record the history of
favored on-link prefixes on the AIL.

It introduces the `otHistoryTrackerFavoredOnLinkPrefix` structure and
a new API, `otHistoryTrackerIterateFavoredOnLinkPrefixHistory()`, to
access the recorded data. The new CLI command `history onlinkprefix`
is added to display the tracked history, showing the prefix, its age,
and whether it is local to the Border Router.