[radio] clarify radio Sleep state
This commit clarifies the Sleep state of radio to be maybe sleep with
the platform supports OT_RADIO_CAPS_RX_ON_WHEN_IDLE.

With this clarification, the Receive() is equivalent to say the radio
should never be turned off, and Sleep() is equivalent to say the radio
may be turned off when idle. This clarification simplifies the semantic
of radio API to support sleepy need with scheduled activities.

This commit also updates the sub mac layer interface to align with this
behavior.